Demo/Website/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
urllib3==1.26.15
